Move types to setup-ganache
